+++ sff/structs.py
@@ -57,8 +57,6 @@ class MainMenu(Enum):
     REMOVE_DRM = "Remove SteamStub DRM (Steamless)"
     DL_USER_GAME_STATS = "Download UserGameStatsSchema (achievements w/o gbe_fork)"
     MULTIPLAYER_FIX = "Apply multiplayer fix (online-fix.me)"
-    CREAMAPI_MP_FIX = "Apply CreamAPI Multiplayer Fix"
-    CREAMAPI_MP_RESTORE = "Restore CreamAPI Multiplayer Fix"
     RYUU_FIX = "Fixes/bypasses (generator.ryuu.lol)"
     OFFLINE_FIX = "Offline Mode Fix"
     if sys.platform == "win32":
@@ -88,8 +86,6 @@ GameSpecificChoices = Literal[
     MainMenu.DL_WORKSHOP_ITEM,
     MainMenu.CHECK_MOD_UPDATES,
     MainMenu.MULTIPLAYER_FIX,
-    MainMenu.CREAMAPI_MP_FIX,
-    MainMenu.CREAMAPI_MP_RESTORE,
     MainMenu.RYUU_FIX,
     MainMenu.MANAGE_DLC_UNLOCKERS
 ]
@@ -102,8 +98,6 @@ GAME_SPECIFIC_CHOICES = (
     MainMenu.DL_WORKSHOP_ITEM,
     MainMenu.CHECK_MOD_UPDATES,
     MainMenu.MULTIPLAYER_FIX,
-    MainMenu.CREAMAPI_MP_FIX,
-    MainMenu.CREAMAPI_MP_RESTORE,
     MainMenu.RYUU_FIX,
     MainMenu.MANAGE_DLC_UNLOCKERS
 )